feat: core

+++ setup/stages/00001-enable-display.sh
@@ -5,7 +5,9 @@ source /etc/tinybox-release
 
 echo "atext,Preparing.. ,Preparing ..,Preparing. ." | nc -U /run/tinybox-screen.sock
 
-systemctl enable tinybox-display
-systemctl enable tinybox-button
-systemctl enable tinybox-poweroff
-systemctl enable tinybox-reboot
+if [[ -z "$TINYGRAD_CORE" ]]; then
+  systemctl enable tinybox-display
+  systemctl enable tinybox-button
+  systemctl enable tinybox-poweroff
+  systemctl enable tinybox-reboot
+fi
